Clean up CommentTypes classes:
- Use TextFileKeywordDefinition structs instead of keywordPairs string lists.
- Rename TextFileKeywords to TextFileKeywordGroup.
- Replace DuplicateWithoutKeywords() with just Duplicate().  A little less memory efficient, but more generalized and less weird.
- Return ILists for properties instead of Lists.
- Add _Stage2 to Manager.Start functions that should only be called in stage 2 because they depend on Languages.Manager.
- Fix some links in the comments.

+++ Engine/Source/CommentTypes/ConfigFiles/BinaryFileParser.cs
@@ -203,7 +203,7 @@ namespace CodeClear.NaturalDocs.Engine.CommentTypes.ConfigFiles
 				foreach (var keywordDefinition in config.KeywordDefinitions)
 					{
 					file.WriteString( keywordDefinition.Keyword );
-					file.WriteByte( (keywordDefinition.Plural ? 1 : 0) );
+					file.WriteByte( (byte)(keywordDefinition.Plural ? 1 : 0) );
 					file.WriteInt32( keywordDefinition.CommentTypeID );
 					file.WriteInt32( (keywordDefinition.IsLanguageSpecific ? keywordDefinition.LanguageID : 0) );
 					}
